refactor: move graphql playlist validators
