+++ app/GraphQL/Validators/Mutation/List/Playlist/Track/CreatePlaylistTrackValidator.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace App\GraphQL\Validators\Mutation\List\Playlist;
+namespace App\GraphQL\Validators\Mutation\List\Playlist\Track;
 
 use App\Contracts\Models\HasHashids;
 use App\Models\List\Playlist;
@@ -15,9 +15,9 @@ use Illuminate\Validation\Rule;
 use Nuwave\Lighthouse\Validation\Validator;
 
 /**
- * Class UpdatePlaylistTrackValidator.
+ * Class CreatePlaylistTrackValidator.
  */
-class UpdatePlaylistTrackValidator extends Validator
+class CreatePlaylistTrackValidator extends Validator
 {
     /**
      * Specify validation rules for the arguments.
@@ -27,26 +27,15 @@ class UpdatePlaylistTrackValidator extends Validator
     public function rules(): array
     {
         $playlistHashid = $this->arg('playlist');
-        $hashid = $this->arg(PlaylistTrack::ATTRIBUTE_HASHID);
-        $entryId = $this->arg(PlaylistTrack::ATTRIBUTE_ENTRY);
-        $videoId = $this->arg(PlaylistTrack::ATTRIBUTE_VIDEO);
-
         $playlist = Playlist::query()->firstWhere(Playlist::ATTRIBUTE_HASHID, $playlistHashid);
-        $track = PlaylistTrack::query()->firstWhere(PlaylistTrack::ATTRIBUTE_HASHID, $hashid);
 
         return [
             PlaylistTrack::ATTRIBUTE_ENTRY => [
-                'sometimes',
                 'required',
                 'integer',
                 Rule::exists(AnimeThemeEntry::class, AnimeThemeEntry::ATTRIBUTE_ID),
-                Rule::when(
-                    ! empty($videoId),
-                    [
-                        Rule::exists(AnimeThemeEntryVideo::class, AnimeThemeEntryVideo::ATTRIBUTE_ENTRY)
-                            ->where(AnimeThemeEntryVideo::ATTRIBUTE_VIDEO, $videoId),
-                    ]
-                ),
+                Rule::exists(AnimeThemeEntryVideo::class, AnimeThemeEntryVideo::ATTRIBUTE_ENTRY)
+                    ->where(AnimeThemeEntryVideo::ATTRIBUTE_VIDEO, $this->arg(PlaylistTrack::ATTRIBUTE_VIDEO)),
             ],
             PlaylistTrack::RELATION_NEXT => [
                 'sometimes',
@@ -54,8 +43,7 @@ class UpdatePlaylistTrackValidator extends Validator
                 'string',
                 Str::of('prohibits:')->append(PlaylistTrack::RELATION_PREVIOUS)->__toString(),
                 Rule::exists(PlaylistTrack::class, HasHashids::ATTRIBUTE_HASHID)
-                    ->where(PlaylistTrack::ATTRIBUTE_PLAYLIST, $playlist?->getKey())
-                    ->whereNot(PlaylistTrack::ATTRIBUTE_ID, $track?->getKey()),
+                    ->where(PlaylistTrack::ATTRIBUTE_PLAYLIST, $playlist?->getKey()),
             ],
             PlaylistTrack::RELATION_PREVIOUS => [
                 'sometimes',
@@ -63,21 +51,14 @@ class UpdatePlaylistTrackValidator extends Validator
                 'string',
                 Str::of('prohibits:')->append(PlaylistTrack::RELATION_NEXT)->__toString(),
                 Rule::exists(PlaylistTrack::class, HasHashids::ATTRIBUTE_HASHID)
-                    ->where(PlaylistTrack::ATTRIBUTE_PLAYLIST, $playlist?->getKey())
-                    ->whereNot(PlaylistTrack::ATTRIBUTE_ID, $track?->getKey()),
+                    ->where(PlaylistTrack::ATTRIBUTE_PLAYLIST, $playlist?->getKey()),
             ],
             PlaylistTrack::ATTRIBUTE_VIDEO => [
-                'sometimes',
                 'required',
                 'integer',
                 Rule::exists(Video::class, Video::ATTRIBUTE_ID),
-                Rule::when(
-                    ! empty($entryId),
-                    [
-                        Rule::exists(AnimeThemeEntryVideo::class, AnimeThemeEntryVideo::ATTRIBUTE_VIDEO)
-                            ->where(AnimeThemeEntryVideo::ATTRIBUTE_ENTRY, $entryId),
-                    ]
-                ),
+                Rule::exists(AnimeThemeEntryVideo::class, AnimeThemeEntryVideo::ATTRIBUTE_VIDEO)
+                    ->where(AnimeThemeEntryVideo::ATTRIBUTE_ENTRY, $this->arg(PlaylistTrack::ATTRIBUTE_ENTRY)),
             ],
         ];
     }
